refactor(form-fields): extract ComboBoxField, rename the directory to Form
Sixteen selectors each rebuilt the same envelope by hand — the layout root, a
`slot='label'` Label, a `slot='input'` control, the id wiring, and the hidden
label's `aria-label`. Five chances per file to get it wrong, which is how
TripPurposeComboBox came to be missing its control slot.

`ComboBoxField` owns the envelope and hands the control its props, so no caller
writes a slot name. It takes the control as a render prop rather than proxying
ComboBox's props: `SingleSelectComboBoxProps` embeds `OneOf<[{options},
{groups}]>`, and an `Omit` over that collapses the union so neither key stays
required. The render prop also fits the creatable and custom-chrome selectors,
which a fixed prop surface would not.

Two selectors hardcoded their input id (`'normality'`, `'type'`) and a third
defaulted to one (`'account_name'`), so two on a page produced duplicate ids
and a label pointing at the wrong control. Ids are generated now, with an
override for the one caller that renders its own label outside the field.

Also renames `blocks/forms` to `blocks/Form`, matching the PascalCase every
other component directory uses, and the story titles with it.
