+++ src/components/blocks/Form/FormDatePickerField.stories.tsx
@@ -1,7 +1,10 @@
 import { CalendarDate } from '@internationalized/date'
 import { type Meta, type StoryObj } from '@storybook/react-vite'
 
-import { FormDateField, type FormDateFieldProps } from '@blocks/forms/FormDateField'
+import {
+  FormDatePickerField,
+  type FormDatePickerFieldProps,
+} from '@blocks/Form/FormDatePickerField'
 
 import {
   COMMON_FIELD_VARIANTS,
@@ -10,26 +13,30 @@ import {
   FormFieldVariantGallery,
 } from '@test-utils/storybook/formField'
 
-const LABEL = 'Trip date'
+const LABEL = 'Effective date'
 const VALUE = new CalendarDate(2026, 3, 14)
 
-type Variant = FormFieldVariant<CalendarDate | null, FormDateFieldProps>
+type Variant = FormFieldVariant<CalendarDate | null, FormDatePickerFieldProps>
 
 const VARIANTS: ReadonlyArray<Variant> = [
   ...COMMON_FIELD_VARIANTS,
   ...ERROR_FIELD_VARIANTS,
   { label: 'empty', value: null },
+  {
+    label: 'bounded to March 2026',
+    props: { minDate: new CalendarDate(2026, 3, 1), maxDate: new CalendarDate(2026, 3, 31) },
+  },
 ]
 
-const meta: Meta<typeof FormDateField> = {
-  title: 'Blocks/Forms/FormDateField',
-  component: FormDateField,
+const meta: Meta<typeof FormDatePickerField> = {
+  title: 'Blocks/Form/FormDatePickerField',
+  component: FormDatePickerField,
   args: { label: LABEL },
 }
 
 export default meta
 
-type Story = StoryObj<typeof FormDateField>
+type Story = StoryObj<typeof FormDatePickerField>
 
 export const AllVariants: Story = {
   parameters: { chromatic: { viewports: [1280] } },
@@ -37,7 +44,7 @@ export const AllVariants: Story = {
     <FormFieldVariantGallery
       defaultValue={VALUE}
       variants={VARIANTS}
-      renderField={props => <FormDateField<CalendarDate> label={LABEL} {...props} />}
+      renderField={props => <FormDatePickerField<CalendarDate> label={LABEL} {...props} />}
     />
   ),
 }
