+++ src/components/blocks/Form/FormNumberField.tsx
@@ -4,8 +4,8 @@ import { useFieldContext } from '@hooks/features/forms/useForm'
 import { Input } from '@ui/Input/Input'
 import { InputGroup } from '@ui/Input/InputGroup'
 import { NumberField } from '@ui/NumberField/NumberField'
-import { FormFieldShell, useFormField } from '@blocks/forms/FormFieldShell'
-import type { CommonFormFieldProps } from '@blocks/forms/types'
+import { FormFieldShell, useFormField } from '@blocks/Form/FormFieldShell'
+import type { CommonFormFieldProps } from '@blocks/Form/types'
 
 export type FormNumberFieldProps = CommonFormFieldProps & {
   minValue?: number
