+++ src/components/blocks/Form/FormSwitchField.stories.tsx
@@ -1,6 +1,7 @@
 import { type Meta, type StoryObj } from '@storybook/react-vite'
+import { Sparkles } from 'lucide-react'
 
-import { FormCheckboxField, type FormCheckboxFieldProps } from '@blocks/forms/FormCheckboxField'
+import { FormSwitchField, type FormSwitchFieldProps } from '@blocks/Form/FormSwitchField'
 
 import {
   COMMON_FIELD_VARIANTS,
@@ -9,25 +10,26 @@ import {
   FormFieldVariantGallery,
 } from '@test-utils/storybook/formField'
 
-const LABEL = 'Exclude from reports'
+const LABEL = 'Auto-categorize transactions'
 
-type Variant = FormFieldVariant<boolean, FormCheckboxFieldProps>
+type Variant = FormFieldVariant<boolean, FormSwitchFieldProps>
 
 const VARIANTS: ReadonlyArray<Variant> = [
   ...COMMON_FIELD_VARIANTS,
   ...ERROR_FIELD_VARIANTS,
   { label: 'unselected', value: false },
+  { label: 'label icon', props: { slots: { LabelIcon: <Sparkles size={14} /> } } },
 ]
 
-const meta: Meta<typeof FormCheckboxField> = {
-  title: 'Blocks/Forms/FormCheckboxField',
-  component: FormCheckboxField,
+const meta: Meta<typeof FormSwitchField> = {
+  title: 'Blocks/Form/FormSwitchField',
+  component: FormSwitchField,
   args: { label: LABEL },
 }
 
 export default meta
 
-type Story = StoryObj<typeof FormCheckboxField>
+type Story = StoryObj<typeof FormSwitchField>
 
 export const AllVariants: Story = {
   parameters: { chromatic: { viewports: [1280] } },
@@ -35,7 +37,7 @@ export const AllVariants: Story = {
     <FormFieldVariantGallery
       defaultValue
       variants={VARIANTS}
-      renderField={props => <FormCheckboxField label={LABEL} {...props} />}
+      renderField={props => <FormSwitchField label={LABEL} {...props} />}
     />
   ),
 }
