+++ src/components/features/categorization/CategorizationRuleForm/CategorizationRuleForm.tsx
@@ -9,7 +9,7 @@ import { useIntlFormatter } from '@hooks/utils/i18n/useIntlFormatter'
 import { SubmitButton } from '@ui/Button/SubmitButton'
 import { Form } from '@ui/Form/Form'
 import { HStack, VStack } from '@ui/Stack/Stack'
-import { FieldErrors } from '@blocks/forms/FieldErrors'
+import { FieldErrors } from '@blocks/Form/FieldErrors'
 import { CategorySelect } from '@features/categorization/CategorizationRuleForm/CategorySelect'
 import { CounterpartySelect } from '@features/categorization/CategorizationRuleForm/CounterpartySelect'
 import { type CategorizationRuleFormState, type DirectionFormValue, getRuleTransactionDescription } from '@features/categorization/CategorizationRuleForm/formUtils'
