+++ src/components/features/customerVendor/VendorSelector/VendorSelector.tsx
@@ -1,4 +1,4 @@
-import { useCallback, useId, useMemo } from 'react'
+import { useCallback, useMemo } from 'react'
 import { useTranslation } from 'react-i18next'
 
 import { type Vendor } from '@schemas/vendor'
@@ -6,9 +6,8 @@ import { getVendorName } from '@utils/vendor'
 import { useGetListVendors } from '@api/businesses/[business-id]/vendors/get'
 import { useDebouncedSearchInput } from '@hooks/utils/debouncing/useDebouncedSearchQuery'
 import { ComboBox } from '@ui/ComboBox/ComboBox'
-import { VStack } from '@ui/Stack/Stack'
-import { Label, P } from '@ui/Typography/Text'
-import { formFieldLayoutProps } from '@blocks/forms/FormFieldShell'
+import { P } from '@ui/Typography/Text'
+import { ComboBoxField } from '@blocks/Form/ComboBoxField'
 import { VendorAsOption } from '@features/customerVendor/VendorSelector/VendorAsOption'
 
 type VendorSelectorProps = {
@@ -39,7 +38,6 @@ export function VendorSelector({
 }: VendorSelectorProps) {
   const { t } = useTranslation()
   const resolvedLabel = label ?? t('customerVendor:label.vendor', 'Vendor')
-  const layoutProps = formFieldLayoutProps({ className, inline, showLabel })
 
   const { searchQuery, handleInputChange } = useDebouncedSearchInput({
     initialInputState: () => '',
@@ -77,32 +75,30 @@ export function VendorSelector({
     [options, handleInputChange, selectedVendor, onSelectedVendorChange],
   )
 
-  const inputId = useId()
   const isLoadingWithoutFallback = isLoading && !flattenedData
 
   return (
-    <VStack {...layoutProps}>
-      {showLabel && <Label slot='label' htmlFor={inputId} size='sm'>{resolvedLabel}</Label>}
-      <ComboBox
-        slot='input'
-        options={options}
-        selectedValue={selectedVendorForComboBox}
-        onSelectedValueChange={handleSelectionChange}
-        onInputValueChange={handleInputChange}
-        inputId={inputId}
-        placeholder={placeholder}
-        isDisabled={isLoadingWithoutFallback || isError}
-        isError={isError}
-        isInvalid={isInvalid}
-        isLoading={isLoadingWithoutFallback}
-        isReadOnly={isReadOnly}
-        isClearable
-        slots={{
-          EmptyMessage: <P variant='subtle'>{t('customerVendor:empty.matching_vendors', 'No matching vendors')}</P>,
-          ErrorMessage: t('customerVendor:error.load_vendors', 'An error occurred while loading vendors.'),
-        }}
-        aria-label={showLabel ? undefined : resolvedLabel}
-      />
-    </VStack>
+    <ComboBoxField label={resolvedLabel} className={className} inline={inline} showLabel={showLabel}>
+      {controlProps => (
+        <ComboBox
+          {...controlProps}
+          options={options}
+          selectedValue={selectedVendorForComboBox}
+          onSelectedValueChange={handleSelectionChange}
+          onInputValueChange={handleInputChange}
+          placeholder={placeholder}
+          isDisabled={isLoadingWithoutFallback || isError}
+          isError={isError}
+          isInvalid={isInvalid}
+          isLoading={isLoadingWithoutFallback}
+          isReadOnly={isReadOnly}
+          isClearable
+          slots={{
+            EmptyMessage: <P variant='subtle'>{t('customerVendor:empty.matching_vendors', 'No matching vendors')}</P>,
+            ErrorMessage: t('customerVendor:error.load_vendors', 'An error occurred while loading vendors.'),
+          }}
+        />
+      )}
+    </ComboBoxField>
   )
 }
