+++ src/components/features/generalLedger/AccountSubtypeComboBox/AccountSubtypeComboBox.tsx
@@ -4,9 +4,7 @@ import { useTranslation } from 'react-i18next'
 import { type LedgerAccountType } from '@schemas/generalLedger/ledgerAccount'
 import { ComboBox } from '@ui/ComboBox/ComboBox'
 import type { ComboBoxOption } from '@ui/ComboBox/types'
-import { HStack } from '@ui/Stack/Stack'
-import { Label } from '@ui/Typography/Text'
-import { formFieldLayoutProps } from '@blocks/forms/FormFieldShell'
+import { ComboBoxField } from '@blocks/Form/ComboBoxField'
 import { SUBTYPES_CONFIG_BY_TYPE } from '@features/generalLedger/constants'
 
 type AccountSubtypeComboBoxProps = {
@@ -29,18 +27,18 @@ export const AccountSubtypeComboBox = ({ label, type, value, onChange, error, in
   }, [t, type])
 
   return (
-    <HStack {...formFieldLayoutProps({ inline })}>
-      <Label slot='label' size='sm' htmlFor='subType'>{label}</Label>
-      <ComboBox
-        slot='input'
-        inputId='subType'
-        options={options}
-        selectedValue={options.find(option => option.value === value) ?? null}
-        onSelectedValueChange={option => onChange(option?.value ?? null)}
-        placeholder={t('chartOfAccounts:placeholder.select_sub_type', 'Select a sub-type')}
-        isError={Boolean(error)}
-        slots={{ ErrorMessage: error }}
-      />
-    </HStack>
+    <ComboBoxField label={label} inline={inline}>
+      {controlProps => (
+        <ComboBox
+          {...controlProps}
+          options={options}
+          selectedValue={options.find(option => option.value === value) ?? null}
+          onSelectedValueChange={option => onChange(option?.value ?? null)}
+          placeholder={t('chartOfAccounts:placeholder.select_sub_type', 'Select a sub-type')}
+          isError={Boolean(error)}
+          slots={{ ErrorMessage: error }}
+        />
+      )}
+    </ComboBoxField>
   )
 }
