+++ src/components/features/generalLedger/ParentAccountComboBox/ParentAccountComboBox.tsx
@@ -4,9 +4,7 @@ import { useTranslation } from 'react-i18next'
 import { type LedgerBalancesSchemaType } from '@schemas/generalLedger/ledgerAccount'
 import { ComboBox } from '@ui/ComboBox/ComboBox'
 import type { ComboBoxOption } from '@ui/ComboBox/types'
-import { HStack } from '@ui/Stack/Stack'
-import { Label } from '@ui/Typography/Text'
-import { formFieldLayoutProps } from '@blocks/forms/FormFieldShell'
+import { ComboBoxField } from '@blocks/Form/ComboBoxField'
 import { useParentAccountOptions } from '@features/generalLedger/ParentAccountComboBox/useParentAccountOptions'
 
 type ParentAccountComboBoxProps = {
@@ -27,18 +25,18 @@ export const ParentAccountComboBox = ({ label, data, value, onChange, error, inl
   )
 
   return (
-    <HStack {...formFieldLayoutProps({ inline })}>
-      <Label slot='label' size='sm' htmlFor='parent'>{label}</Label>
-      <ComboBox
-        slot='input'
-        inputId='parent'
-        options={options}
-        selectedValue={options.find(option => option.value === value) ?? null}
-        onSelectedValueChange={option => onChange(option?.value ?? null)}
-        placeholder={t('chartOfAccounts:placeholder.select_parent', 'Select a parent account')}
-        isError={Boolean(error)}
-        slots={{ ErrorMessage: error }}
-      />
-    </HStack>
+    <ComboBoxField label={label} inline={inline}>
+      {controlProps => (
+        <ComboBox
+          {...controlProps}
+          options={options}
+          selectedValue={options.find(option => option.value === value) ?? null}
+          onSelectedValueChange={option => onChange(option?.value ?? null)}
+          placeholder={t('chartOfAccounts:placeholder.select_parent', 'Select a parent account')}
+          isError={Boolean(error)}
+          slots={{ ErrorMessage: error }}
+        />
+      )}
+    </ComboBoxField>
   )
 }
