+++ src/components/features/invoices/InvoiceTermsComboBox/InvoiceTermsComboBox.tsx
@@ -1,4 +1,4 @@
-import { useCallback, useId, useMemo } from 'react'
+import { useCallback, useMemo } from 'react'
 import { type ZonedDateTime } from '@internationalized/date'
 import { differenceInDays, startOfDay } from 'date-fns'
 import { useTranslation } from 'react-i18next'
@@ -6,9 +6,7 @@ import { useTranslation } from 'react-i18next'
 import { InvoiceTermsValues } from '@schemas/invoices/invoiceTerms'
 import { translationKey } from '@utils/i18n/translationKey'
 import { ComboBox } from '@ui/ComboBox/ComboBox'
-import { VStack } from '@ui/Stack/Stack'
-import { Label } from '@ui/Typography/Text'
-import { formFieldLayoutProps } from '@blocks/forms/FormFieldShell'
+import { ComboBoxField } from '@blocks/Form/ComboBoxField'
 
 type InvoiceTermsOption = {
   label: string
@@ -89,23 +87,19 @@ export const InvoiceTermsComboBox = ({ value, onValueChange, isReadOnly }: Invoi
   }
   const onSelectedValueChange = useCallback(handleChange, [onValueChange])
 
-  const inputId = useId()
-
   return (
-    <VStack {...formFieldLayoutProps({ className: 'Layer__InvoiceForm__Field__Terms', inline: true })}>
-      <Label slot='label' size='sm' htmlFor={inputId}>
-        {t('invoices:label.terms', 'Terms')}
-      </Label>
-      <ComboBox
-        slot='input'
-        options={options}
-        onSelectedValueChange={onSelectedValueChange}
-        selectedValue={selectedOption}
-        isSearchable={false}
-        isClearable={false}
-        inputId={inputId}
-        isReadOnly={isReadOnly}
-      />
-    </VStack>
+    <ComboBoxField label={t('invoices:label.terms', 'Terms')} className='Layer__InvoiceForm__Field__Terms' inline>
+      {controlProps => (
+        <ComboBox
+          {...controlProps}
+          options={options}
+          onSelectedValueChange={onSelectedValueChange}
+          selectedValue={selectedOption}
+          isSearchable={false}
+          isClearable={false}
+          isReadOnly={isReadOnly}
+        />
+      )}
+    </ComboBoxField>
   )
 }
